every functionality till sidebar and delete file is done now

+++ Backend/FASTAPI/delete_api/delete_project_api.py
@@ -5,6 +5,7 @@ from DATABASE.Tables.users_table import User
 from DATABASE.Tables.projects_table import Project
 from DATABASE.database import get_db
 from FASTAPI.post_apis.Users_api.auth.Signin_api import get_current_user
+from DATABASE.Tables.file_table import File
 
 
 router = APIRouter()
@@ -18,14 +19,21 @@ def delete_project(
     project = db.query(Project).filter(
         and_(
             Project.project_name == project_name,
-            Project.user_id == current_user.id
+            Project.user_id == current_user.id,
         )
     ).first()
 
     if not project:
         raise HTTPException(
             status_code= 404,
-            detail="Project not found toooo"
+            detail="Project not found"
+        )
+    file = db.query(File).filter(File.project_id == project.id).first()
+
+    if file:
+        raise HTTPException(
+            status_code=409,
+            detail="there are files in this project so it can not be delete"
         )
     try:
         db.delete(project)
@@ -34,6 +42,6 @@ def delete_project(
     except Exception:
         db.rollback()
         raise HTTPException(
-            status_code= 404,
+            status_code= 500,
             detail="faild to delete "
         )
